Genericize StringBuilder
